Fix command text not working with permalink and move more modal logic out of main component.
